LayoutTests/platform/ios-sim-deprecated/fast/events/selectstart-by-arrow-keys-expected.txt

This test ensures selectstart is fired when selection is created by arrow keys.

If running this test manually, click on div ("Hello World") element and try to select the text using arrow keys.
Expected result : SelectStart event will fire when user starts extending selection.
Hello World
Initial state: PASS
Check (Right arrow): PASS
Check (Right arrow + Shift): FAIL - expected 1 events but got 0 events
Check (Right arrow + Shift + Control): FAIL - expected 1 events but got 0 events
Check (End + Shift): FAIL - expected 1 events but got 0 events
Check (Home + Shift): FAIL - expected 1 events but got 0 events
Check (End + Shift): FAIL - expected 1 events but got 0 events
Check (Left arrow): FAIL - expected 1 events but got 0 events
Check (LeftArrow + Shift + Control): FAIL - expected 2 events but got 0 events
Check (Left arrow + Shift): FAIL - expected 2 events but got 0 events
Check (Home + Shift): FAIL - expected 2 events but got 0 events
Check (Home + Control): FAIL - expected 2 events but got 0 events
Done.
